% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishpals_colors.R
\docType{data}
\name{cfs}
\alias{cfs}
\title{Cramer Fish Sciences Palette}
\format{An object of class \code{character} of length 5.}
\usage{
cfs
}
\description{
Cramer Fish Sciences Palette
}
\keyword{internal}
