% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishpals-package.R
\name{fishpals-package}
\alias{fishpals-package}
\alias{fishpals}
\title{The 'fishpals' package}
\description{
Heavily influenced by Diane Cook's `ochre` package and Karthik
Ram's `wesanderson` package, this package
provides species-specific color palettes (generated from pictures of fish 
that 'Cramer Fish Sciences' has worked with in the past), and some
lightweight ggplot2 themes.
}
