% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishpals-package.R
\docType{package}
\name{fishpals}
\alias{fishpals}
\alias{fishpals-package}
\title{Species-specific Fish Color Palettes for Plots}
\description{
Modeled after Karthik Ram's \href{https://CRAN.R-project.org/package=wesanderson}{wesanderson}
palettes package, methods areprovided to create species-specific color palettes generated
from pictures of fish that \href{http://www.fishsciences.net/}{Cramer Fish Sciences} (CFS)
has worked with in the past. Since CFS works mostly with North American, northwestern
fish species, we're talking about a lot of greens, browns, and grays. But we're open
to adding some tropical palettes.
}
\author{
Myfanwy Johnston \href{mailto:myfanwy.johnston@fishsciences.net}{myfanwy.johnston@fishsciences.net}
}
