% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishpals_pals.R, R/get_fishpals_cols.R
\name{get_fishpals_cols}
\alias{get_fishpals_cols}
\title{Fish palettes - groups colors into their palettes
The name of the palette is set here}
\usage{
get_fishpals_cols(...)

get_fishpals_cols(...)
}
\arguments{
\item{...}{Quoted color name or vector of names.}
}
\value{
The 6-digit hexes attached to the name
}
\description{
Fish palettes - groups colors into their palettes
The name of the palette is set here

Function that extracts hex codes form the fishpals_colors vector by name
}
\examples{
get_fishpals_cols()
get_fishpals_cols("chinook red")
}
