% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpals.R
\name{getpals}
\alias{getpals}
\title{Function prints the names of all the available palettes in fishpals}
\usage{
getpals()
}
\value{
The names of the fishpals palettes
}
\description{
Function prints the names of all the available palettes in fishpals
}
\examples{
getpals()
}
