% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_foundationfp}
\alias{theme_foundationfp}
\title{Fishpals ggplot2 themes}
\usage{
theme_foundationfp(base_size = 11, base_family = "Geneva")
}
\description{
lightweight ggplot2 themes for Cramer Fish Sciences and fish lovers alike

\describe{
\item{theme_foundation}{Code modified from [jrnold's theme_foundation](https://github.com/jrnold/ggthemes/blob/master/R/theme-foundation.R); designed to be a foundation from which the other themes are built, and not meant to be used directly.}
\item{theme_report}{has gridlines}
\item{theme_pres}{intended for use in presentations. Larger base_font}
\item{theme_pub}{good template for publications}
}
}
\examples{
library(ggplot2)

extrafont::loadfonts(quiet=TRUE)

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) +
  scale_color_fishpals("genidaqs") +
  theme_report()

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
 geom_point(size = 4, alpha = .6) +
 scale_color_fishpals(discrete = FALSE, palette = "genidaqs") +
 theme_pub()

}
