% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_report}
\alias{theme_report}
\alias{theme_pres}
\alias{theme_pub}
\title{Fishpals ggplot2 themes}
\usage{
theme_report()

theme_pres()

theme_pub()
}
\description{
lightweight ggplot2 themes for Cramer Fish Sciences and fish lovers alike

\describe{
\item{theme_report}{has gridlines}
\item{theme_pres}{intended for use in presentations. Larger base_font}
\item{theme_pub}{good template for publications}
}
}
\examples{
library(ggplot2)

extrafont::loadfonts(quiet=TRUE)

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) +
  scale_color_fishpals("genidaqs") +
  theme_report()

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
 geom_point(size = 4, alpha = .6) +
 scale_color_fishpals(discrete = FALSE, palette = "genidaqs") +
 theme_pub()

}
