% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewpals.R
\name{viewpals}
\alias{viewpals}
\title{Show colors associated with a palette - adapted from the [`scales` package](https://cran.r-project.org/web/packages/scales/index.html) for use with fishpals}
\usage{
viewpals(fishpal, labels=TRUE, borders = NULL, cex_label = 1)
}
\arguments{
\item{fishpal}{a string matching one of the fishpals color palettes (see fishpals::fish_pals for the full list); default is "CFS"}

\item{labels}{boolean, whether to show the hexadecimal representation of the colours in each tile}

\item{borders}{colour of the borders of the tiles; matches the border argument of `graphics::rect()`. The default means par("fg"). Use border = NA to omit borders.}

\item{cex_label}{size of printed labels, works the same as cex parameter of plot()}
}
\description{
A quick and dirty way to view fishpalettes
}
\examples{
viewpals("genidaqs")
viewpals("adultchinook")
}
