% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishpals_colors.R
\docType{data}
\name{fish_pals}
\alias{fish_pals}
\title{Fish palettes Core Colors list - still growing}
\format{An object of class \code{list} of length 3.}
\usage{
fish_pals
}
\description{
Fish palettes Core Colors list - still growing
}
\keyword{internal}
