% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_report}
\alias{theme_report}
\alias{theme_pres}
\alias{theme_pub}
\title{Fishpals ggplot2 themes}
\usage{
theme_report(
  inner_border = TRUE,
  outer_border = TRUE,
  outer_border_color = "gray40",
  inner_border_color = "gray25",
  base_size = 11,
  base_family = "Geneva"
)

theme_pres(transparent = TRUE, base_size = 14, base_family = "Helvetica")

theme_pub(transparent = TRUE, base_size = 10.5, base_family = "Arial")
}
\description{
lightweight ggplot2 themes for Cramer Fish Sciences and fish lovers alike
}
\details{
\describe{
\item{theme_report}{good template theme for reports. Default inner panel border and light gridlines}
\item{theme_pres}{intended for use in presentations. Larger default fonts, no gridlines, clean background suited for transparency options}
\item{theme_pub}{good baseline template for publications.  Smaller default fonts, theme suited for black and white.}
}
}
\examples{
library(ggplot2)

extrafont::loadfonts(quiet=TRUE)

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) +
  scale_color_fishpals("genidaqs") +
  theme_report(inner_border = TRUE)

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
   geom_point(size = 4, alpha = .6) +
   scale_color_fishpals(discrete = FALSE, palette = "genidaqs") +
   theme_pres(base_family = "Megrim")

}
