% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewpals.R
\name{viewpals}
\alias{viewpals}
\title{Quick preview of the core colors in each fish palette from \code{fishpals()}.}
\usage{
viewpals(palette)
}
\arguments{
\item{palette}{The bare/unquoted name or string of the color palette you want
to use from the fishpals_colors list.
Options are \code{juvsteel} for juvenile steelhead, \code{adultsteel} for adult
steelhead, and \code{adultchinook} for adult Chinook salmon.}
}
\description{
Quick preview of the core colors in each fish palette from \code{fishpals()}.
}
\examples{
viewpals(adultsteel)
viewpals("adultsteel")
viewpals(juvsteel)
viewpals(adultchinook)
}
