% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewpals.R
\name{viewpals}
\alias{viewpals}
\title{Display colors associated with a palette}
\usage{
viewpals(fishpal = "cfs", labels = TRUE, borders = NULL, cex_label = 0.75)
}
\arguments{
\item{fishpal}{a string matching one of the \code{fishpals} color palettes (see \code{fishpals::fish_pals} for the full list).  The default is "cfs"}

\item{labels}{boolean, whether to show the hexadecimal representation of the colours in each tile}

\item{borders}{color of the borders of the tiles; matches the border argument of \code{graphics::rect()}. The default means par("fg"). Use border = NA to omit borders.}

\item{cex_label}{size of printed labels, works the same as \code{cex} parameter of plot()}
}
\description{
A quick way to view \code{fishpals}' color palettes
}
\details{
Adapted from the \code{scales} package (\url{https://cran.r-project.org/web/packages/scales/index.html}) for use with fishpals
}
\examples{
viewpals("genidaqs")
viewpals("adultchinook")
}
